% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospectorMahery-package.R
\docType{package}
\name{prospectorMahery-package}
\alias{prospectorMahery}
\alias{prospectorMahery-package}
\title{prospectorMahery: Prospector for MAHERY Project Data}
\description{
Dataset prospector to generate the MAHERY data set related to the project "Cohort Profile: The Madagascar Health and Environmental Research (MAHERY) study in north-eastern Madagascar".
}
\author{
\strong{Maintainer}: Tinashe Michael Tapera \email{ttapera@hsph.harvard.edu} (\href{https://orcid.org/0000-0001-9080-5010}{ORCID})

Authors:
\itemize{
  \item Tinashe Michael Tapera \email{ttapera@hsph.harvard.edu} (\href{https://orcid.org/0000-0001-9080-5010}{ORCID})
  \item Stephanie Wu \email{swu@g.harvard.edu} (\href{https://orcid.org/0000-0002-2258-7493}{ORCID}) [contributor, data contributor]
}

Other contributors:
\itemize{
  \item Christopher Golden \email{golden@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-2258-7493}{ORCID}) [copyright holder]
}

}
\keyword{internal}
