% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_inputs.R
\name{register_inputs}
\alias{register_inputs}
\title{List matching file paths from a zip archive}
\usage{
register_inputs(input_path, regex = "MAHERY")
}
\arguments{
\item{input_path}{path to the zipped dataset}

\item{regex}{regular expression to filter the files to be registered}
}
\value{
vector of file paths
}
\description{
Lists the contents of a zipped dataset and returns the paths of files
whose names match the supplied regular expression.
}
\examples{
register_inputs(
  here(
    "data", 
    "inputs", 
    "01_gold_mine", 
    "madagascar-cohort-data-zip/madagascar_data_cleaning-main.zip"
    ),
  regex = "Data\\\\/MAHERY"
  )
}
